Move useIsLoggedIn to hooks folder inside module; set optional to false for Stacks and Params
