+++ src/modules/city-pass/Account.tsx
@@ -4,8 +4,8 @@ import {DigiDButton} from '@/components/ui/buttons/DigiDButton'
 import {Column} from '@/components/ui/layout/Column'
 import {Paragraph} from '@/components/ui/text/Paragraph'
 import {useNavigation} from '@/hooks/navigation/useNavigation'
+import {useIsLoggedIn} from '@/modules/city-pass/hooks/useIsLoggedIn'
 import {CityPassRouteName} from '@/modules/city-pass/routes'
-import {useIsLoggedIn} from '@/modules/city-pass/useIsLoggedIn'
 import {ModuleSlug} from '@/modules/slugs'
 import {UserRouteName} from '@/modules/user/routes'
