+++ src/modules/mijn-amsterdam/Account.tsx
@@ -15,8 +15,8 @@ import {useNavigation} from '@/hooks/navigation/useNavigation'
 import {useRoute} from '@/hooks/navigation/useRoute'
 import {usePermission} from '@/hooks/permissions/usePermission'
 import {useHandleLoginDeeplink} from '@/modules/mijn-amsterdam/hooks/useHandleLoginDeeplink'
+import {useIsLoggedIn} from '@/modules/mijn-amsterdam/hooks/useIsLoggedIn'
 import {useLoginMijnAmsterdam} from '@/modules/mijn-amsterdam/hooks/useLoginMijnAmsterdam'
-import {useIsLoggedIn} from '@/modules/mijn-amsterdam/useIsLoggedIn'
 import {ModuleSlug} from '@/modules/slugs'
 import {UserRouteName} from '@/modules/user/routes'
 import {Permissions} from '@/types/permissions'
