+++ src/modules/parking/Account.tsx
@@ -5,9 +5,9 @@ import {NavigationButton} from '@/components/ui/buttons/NavigationButton'
 import {Column} from '@/components/ui/layout/Column'
 import {Paragraph} from '@/components/ui/text/Paragraph'
 import {useNavigation} from '@/hooks/navigation/useNavigation'
+import {useIsLoggedIn} from '@/modules/parking/hooks/useIsLoggedIn'
 import {ParkingRouteName} from '@/modules/parking/routes'
 import {useParkingAccounts} from '@/modules/parking/slice'
-import {useIsLoggedIn} from '@/modules/parking/useIsLoggedIn'
 import {ModuleSlug} from '@/modules/slugs'
 
 export const Account = () => {
